+++ src/entities/FileSearchItem.ts
